Timer/Health bar working, temporary only 1 bar but code works so just need to add another bar.
Currently have a frog spawning extra child(?) bug. Trying to fix.

+++ Classes/Frog.cpp
@@ -133,7 +133,7 @@ int Frog::GetHP(void)
 	return HP;
 }
 
-int Frog::GetDamage(void)
+float Frog::GetDamage(void)
 {
 	return damage;
 }
